workers/agent-runtime/requirements.txt
# Agent Runtime Dependencies

# LLM Services
dashscope>=1.14.0
openai>=1.0.0
anthropic>=0.18.0

# Environment Management
python-dotenv>=1.0.0

# JSON Schema Validation
jsonschema>=4.17.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
hypothesis>=6.82.0
